+++ server/middleware/save-thumbnail.middleware.js
